requirements.txt
discord.py>=2.4,<3
httpx>=0.27,<1
aiosqlite>=0.20,<1
python-dotenv>=1.0,<2
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
pytest>=8.0,<9
pytest-asyncio>=0.23,<1
respx>=0.21,<1
pytest-cov>=5.0,<7

